wrap test and use before each
